feat: Introduce a ComponentBasedAlterationEffect
Extracted from #14

This is a common base class for alteration effects based on a single component. It only requires implementations of component _upsert_ and _update_ to fully define an effect (in addition to the component class and the general effect id). _Removing_ the component can also be customized.

The implementation is extracted from WalkSpeedAlterationEffect and can be found in other effect classes such as SwimSpeedAlterationEffect or StunEffect.