feat: Revision of the effect authority system
The authority system is now shared such that it can be injected via `@In` to other systems. Alteration effects can be registered with the system to allow other modules to contribute new effects that are automatically managed by the authority system.

The registered effects are now exposed so that consumers don't have to instantiate effects on their own but can rely on registered effects:

```java
effectSystem.getEffect(AlterationEffects.STUN).applyEffect(instigator, entity);
```

Uses both the EffectIdentifier for parsing effects from delayed action ids as well as the RecalculateEffect event to trigger recalculation after expiration of a single effect instance.

In addition, adds a simple debug command to apply arbitrary effects to the player issuing the command.

TODO: Add JavaDoc.